@@ -25,28 +25,28 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@types/lodash": "4.14.168",
-    "@types/node": "14.14.31",
-    "@types/ramda": "0.27.38",
-    "@types/selenium-webdriver": "4.0.11",
+    "@types/node": "14.14.40",
+    "@types/ramda": "0.27.40",
+    "@types/selenium-webdriver": "4.0.12",
     "@types/semver": "7.3.4",
-    "@types/yargs": "16.0.0",
+    "@types/yargs": "16.0.1",
     "dockerode": "^3.2.1",
     "rimraf": "^3.0.2",
     "ts-node": "9.1.1",
-    "typescript": "4.1.2"
+    "typescript": "4.2.4"
   },
   "dependencies": {
     "axios": "^0.21.1",
-    "chromedriver": "^88.0.0",
+    "chromedriver": "^89.0.0",
     "cross-env": "7.0.3",
     "dot": "1.1.3",
     "jstat": "1.9.4",
-    "lighthouse": "7.2.0",
+    "lighthouse": "7.3.0",
     "lodash": "4.17.21",
-    "npm-check-updates": "11.1.10",
+    "npm-check-updates": "11.4.1",
     "ramda": "0.27.1",
-    "selenium-webdriver": "4.0.0-beta.1",
-    "semver": "7.3.4",
+    "selenium-webdriver": "4.0.0-beta.3",
+    "semver": "7.3.5",
     "yargs": "16.2.0"
   }
 }
