@@ -23,8 +23,7 @@ module.exports = {
 				{
 					loader: 'babel-loader',
 					options: {
-						presets: ['@babel/preset-env', '@babel/preset-react'],
-						plugins: ['@babel/plugin-proposal-class-properties'],
+						presets: ['@babel/preset-env', ['@babel/preset-react', { runtime: 'automatic' }]],
 					}
 				}
 			]
