@@ -64,6 +64,8 @@ export interface FrameworkData {
     keyed: boolean;
     useShadowRoot: boolean;
     useRowShadowRoot: boolean;
+    shadowRootName: string|undefined;
+    buttonsInShadowRoot: boolean;
     issues: number[];
 }
 
@@ -87,7 +89,8 @@ export interface FrameworkId {
 
 abstract class FrameworkVersionInformationValid implements FrameworkId {
     public url: string;
-    constructor(public keyedType: KeyedType, public directory: string, customURL: string|undefined, public useShadowRoot: boolean, public useRowShadowRoot: boolean, public issues: number[]) {
+    constructor(public keyedType: KeyedType, public directory: string, customURL: string|undefined, public useShadowRoot: boolean, public useRowShadowRoot: boolean,       
+      public shadowRootName: string, public buttonsInShadowRoot: boolean, public issues: number[]) {
         this.keyedType = keyedType;
         this.directory = directory;
         this.url = 'frameworks/'+keyedType+'/'+directory + (customURL ? customURL : '');
@@ -96,14 +99,14 @@ abstract class FrameworkVersionInformationValid implements FrameworkId {
 
 export class FrameworkVersionInformationDynamic extends FrameworkVersionInformationValid  {
     constructor(keyedType: KeyedType, directory: string, public packageNames: string[],
-        customURL: string|undefined, useShadowRoot: boolean = false, useRowShadowRoot: boolean = false, issues: number[]) {
-            super(keyedType, directory, customURL, useShadowRoot, useRowShadowRoot, issues);
+        customURL: string|undefined, useShadowRoot: boolean = false, useRowShadowRoot: boolean = false, public shadowRootName: string, public buttonsInShadowRoot: boolean, issues: number[]) {
+            super(keyedType, directory, customURL, useShadowRoot, useRowShadowRoot, shadowRootName, buttonsInShadowRoot, issues);
         }
     }
 
 export class FrameworkVersionInformationStatic extends FrameworkVersionInformationValid  {
-    constructor(keyedType: KeyedType, directory: string, public frameworkVersion: string, customURL: string|undefined, useShadowRoot: boolean = false, useRowShadowRoot: boolean = false, issues: number[]) {
-        super(keyedType, directory, customURL, useShadowRoot, useRowShadowRoot, issues);
+    constructor(keyedType: KeyedType, directory: string, public frameworkVersion: string, customURL: string|undefined, useShadowRoot: boolean = false, useRowShadowRoot: boolean = false, public shadowRootName: string, public buttonsInShadowRoot: boolean, issues: number[]) {
+        super(keyedType, directory, customURL, useShadowRoot, useRowShadowRoot, shadowRootName, buttonsInShadowRoot, issues);
     }
     getFrameworkData(): FrameworkData {
         return {name: this.directory,
@@ -112,6 +115,8 @@ export class FrameworkVersionInformationStatic extends FrameworkVersionInformati
             keyed: this.keyedType === 'keyed',
             useShadowRoot: this.useShadowRoot,
             useRowShadowRoot: this.useRowShadowRoot,
+            shadowRootName: this.shadowRootName,
+            buttonsInShadowRoot: this.buttonsInShadowRoot,
             issues: this.issues
         }
     }
@@ -177,6 +182,8 @@ async function loadFrameworkInfo(pathInFrameworksDir: string): Promise<Framework
                     packageJSON['js-framework-benchmark']['customURL'],
                     packageJSON['js-framework-benchmark']['useShadowRoot'],
                     packageJSON['js-framework-benchmark']['useRowShadowRoot'],
+                    packageJSON['js-framework-benchmark']['shadowRootName'] ?? 'main-element',
+                    packageJSON['js-framework-benchmark']['buttonsInShadowRoot'] ?? true,
                     packageJSON['js-framework-benchmark']['issues']
                 );
             } else if (typeof packageJSON['js-framework-benchmark']['frameworkVersion'] === 'string') {
@@ -185,6 +192,8 @@ async function loadFrameworkInfo(pathInFrameworksDir: string): Promise<Framework
                     packageJSON['js-framework-benchmark']['customURL'],
                     packageJSON['js-framework-benchmark']['useShadowRoot'],
                     packageJSON['js-framework-benchmark']['useRowShadowRoot'],
+                    packageJSON['js-framework-benchmark']['shadowRootName'] ?? 'main-element',
+                    packageJSON['js-framework-benchmark']['buttonsInShadowRoot'] ?? true,
                     packageJSON['js-framework-benchmark']['issues']
                 );
             } else {
@@ -234,6 +243,8 @@ export class PackageVersionInformationResult {
             keyed: this.framework.keyedType === 'keyed',
             useShadowRoot: this.framework.useShadowRoot,
             useRowShadowRoot: this.framework.useRowShadowRoot,
+            shadowRootName: this.framework.shadowRootName,
+            buttonsInShadowRoot: this.framework.buttonsInShadowRoot,
             issues: this.framework.issues
         }
     }
